refactor(notebook): share host shims
