+++ apps/notebook/src/notebook-surface.ts
@@ -6,39 +6,7 @@ export { PresenceValueProvider, type PresenceContextValue } from "./contexts/Pre
 export { CrdtBridgeProvider } from "./hooks/useCrdtBridge";
 export { createNotebookController } from "./lib/notebook-controller";
 export { createNotebookCellId, type NotebookCellIdRandomSource } from "./lib/notebook-cell-id";
-export {
-  applyExecutionViewChangeset,
-  applyOutputChangeset,
-  getOutputProjectionFailures,
-  resetRuntimeStoresProjection,
-  subscribeOutputProjectionFailures,
-  useOutputProjectionFailures,
-} from "./lib/project-runtime-stores";
-export {
-  materializeChangeset,
-  type CellChangeset,
-  type MaterializeDeps,
-} from "./lib/frame-pipeline";
-export {
-  getCellById,
-  getCellIdsSnapshot,
-  getNotebookCellsSnapshot,
-  type NotebookCell,
-} from "./lib/notebook-cells";
-export { shouldPreserveBootstrapProjection } from "@/components/notebook/state/bootstrap-preservation";
-export type { JupyterOutput } from "./types";
+export { materializeChangeset } from "./lib/frame-pipeline";
 export { resetPoolState, setPoolState } from "./lib/pool-state";
-export {
-  resetRuntimeState,
-  runtimeStateStore,
-  setRuntimeState,
-  useRuntimeProjection,
-  useRuntimeState,
-  useRuntimeStateLoaded,
-  useThrottledStatusKey,
-  useWorkstationAttachment,
-} from "@/components/notebook/state/runtime-state";
 export { startCursorDispatch } from "./lib/cursor-registry";
-export { setLoggerHost } from "./lib/logger";
-export { setOpenUrlHost } from "./lib/open-url";
 export { emitBroadcast, emitPresence } from "./lib/notebook-frame-bus";
